Remove broken upgrade wizard
