--- core/src/main/resources/jenkins/install/UpgradeWizard/client-scripts.properties
@@ -1,3 +1,0 @@
-msg.before=Welcome to Jenkins {0}!\u0020
-msg.link=Upgrade now
-msg.after=\u0020to get the new features
