--- core/src/main/resources/jenkins/install/UpgradeWizard/client-scripts_lt.properties
@@ -1,3 +1,0 @@
-msg.before=Jus sveikina Jenkins {0}!\u0020
-msg.link=\u012Ediegti naujas savybes
-msg.after=\u0020kad baigtum\u0117te atnaujinim\u0105
