--- war/src/main/js/templates/upgradeSkippedPanel.hbs
@@ -1,17 +1,0 @@
-<div class="modal-header">
-  <h4 class="modal-title">{{translations.installWizard_upgradeSkipped_title}}</h4>
-</div>
-<div class="modal-body">
-  <div class="jumbotron welcome-panel success-panel">
-    <h1>{{translations.installWizard_upgradeSkipped_banner}}</h1>
-    <p>{{{translations.installWizard_upgradeSkipped_message}}}</p>
-  </div>
-</div>
-<div class="modal-footer">
-  <button type="button" class="btn btn-link install-home">
-    {{translations.installWizard_goBack}}
-  </button>
-  <button type="button" class="btn btn-primary install-done">
-    {{translations.installWizard_skipPluginInstallations}}
-  </button>
-</div>
